cambios para comentarios y cosas opcionales

+++ aspnet-core/src/Servicios.Application.Contracts/Hamburguesa/HamburguesasDTOGet.cs
@@ -10,7 +10,7 @@ namespace Servicios.Domain.Hamburguesa
     public class HamburguesasDTOGet : EntityDto<Guid>
     {
         public string Nombre { get; set; }
-        public decimal? Precio { get; set; }
+        public double? Precio { get; set; }
         public string Descripcion { get; set; }
         public string ImagenUrl { get; set; }
         public List<IngredientesDTO> ListIngredientes { get; set; }
